update-import-excel-view-provided-invoice-provided

+++ database/migrations/2022_05_29_162408_create_provideds_table.php
@@ -15,7 +15,7 @@ class CreateProvidedsTable extends Migration
     {
         Schema::create('provideds', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('tax_code');
+            $table->bigInteger('tax_code');
             $table->string('name',100);
             $table->string('email',100);
             $table->string('phone',10);
